+++ tanaguru-engine/tanaguru-api/src/main/java/com/tanaguru/repository/WebextEngineRepository.java
@@ -1,6 +1,7 @@
 package com.tanaguru.repository;
 
 import java.util.Collection;
+import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -15,6 +16,6 @@ import com.tanaguru.domain.entity.audit.WebextEngine;
 @Repository
 public interface WebextEngineRepository extends JpaRepository<WebextEngine, Long>  {
 
-    Collection<WebextEngine> findAllByEngineVersion(String engineVersion);
+    Optional<WebextEngine> findByEngineVersion(String engineVersion);
     
 }