+++ tanaguru-engine/tanaguru-api/src/main/java/com/tanaguru/service/impl/TestHierarchyServiceImpl.java
@@ -3,11 +3,13 @@ package com.tanaguru.service.impl;
 import com.google.gson.Gson;
 import com.tanaguru.domain.entity.audit.TanaguruTest;
 import com.tanaguru.domain.entity.audit.TestHierarchy;
+import com.tanaguru.domain.entity.audit.WebextEngine;
 import com.tanaguru.domain.jsonmapper.JsonTanaguruWebextTest;
 import com.tanaguru.domain.jsonmapper.JsonTestHierarchy;
 import com.tanaguru.repository.AuditReferenceRepository;
 import com.tanaguru.repository.TanaguruTestRepository;
 import com.tanaguru.repository.TestHierarchyRepository;
+import com.tanaguru.repository.WebextEngineRepository;
 import com.tanaguru.service.TestHierarchyService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -31,7 +33,10 @@ public class TestHierarchyServiceImpl implements TestHierarchyService {
     private final TestHierarchyRepository testHierarchyRepository;
     private final AuditReferenceRepository auditReferenceRepository;
     private final TanaguruTestRepository tanaguruTestRepository;
-
+    private final WebextEngineRepository webextEngineRepository;
+    private final String coreScript;
+    private final String coreScriptVersion;
+    
     @Value("references/wcag-definition.json")
     private ClassPathResource wcagResource;
 
@@ -39,15 +44,33 @@ public class TestHierarchyServiceImpl implements TestHierarchyService {
     private ClassPathResource actTestsPath;
 
     @Autowired
-    public TestHierarchyServiceImpl(TestHierarchyRepository testHierarchyRepository, AuditReferenceRepository auditReferenceRepository, TanaguruTestRepository tanaguruTestRepository) {
+    public TestHierarchyServiceImpl(TestHierarchyRepository testHierarchyRepository, 
+            AuditReferenceRepository auditReferenceRepository, 
+            TanaguruTestRepository tanaguruTestRepository,
+            WebextEngineRepository webextEngineRepository,
+            String coreScript,
+            String coreScriptVersion) {
         this.testHierarchyRepository = testHierarchyRepository;
         this.auditReferenceRepository = auditReferenceRepository;
         this.tanaguruTestRepository = tanaguruTestRepository;
+        this.webextEngineRepository = webextEngineRepository;
+        this.coreScript = coreScript;
+        this.coreScriptVersion = coreScriptVersion;
     }
 
     @PostConstruct
     @Transactional
-    public void insertBaseTestHierarchy() throws IOException {
+    public void insertBaseWebextEngine() throws IOException {
+        if(!webextEngineRepository.findByEngineVersion(coreScriptVersion).isPresent()) {
+            WebextEngine webextEngine = new WebextEngine();
+            webextEngine.setEngineVersion(coreScriptVersion);
+            webextEngine.setEngineContent(coreScript.getBytes());
+            webextEngineRepository.save(webextEngine);
+        }
+        insertBaseTestHierarchy(webextEngineRepository.findByEngineVersion(coreScriptVersion).get());
+    }
+    
+    private void insertBaseTestHierarchy(WebextEngine webextEngine) throws IOException {
         Gson gson = new Gson();
         JsonTestHierarchy wcag = gson.fromJson(
                 StreamUtils.copyToString(
@@ -57,7 +80,7 @@ public class TestHierarchyServiceImpl implements TestHierarchyService {
 
         if (!testHierarchyRepository.findByCodeAndParentIsNull(wcag.getCode()).isPresent()) {
             LOGGER.info("Create reference " + wcag.getCode());
-            TestHierarchy actRef = importTestHierarchy(wcag, null);
+            TestHierarchy actRef = importTestHierarchy(wcag, null, webextEngine);
             String actJson = StreamUtils.copyToString(
                     actTestsPath.getInputStream(),
                     Charset.defaultCharset());
@@ -106,7 +129,7 @@ public class TestHierarchyServiceImpl implements TestHierarchyService {
         }
     }
 
-    private TestHierarchy importTestHierarchy(JsonTestHierarchy jsonTestHierarchy, TestHierarchy parent) {
+    private TestHierarchy importTestHierarchy(JsonTestHierarchy jsonTestHierarchy, TestHierarchy parent, WebextEngine webextEngine) {
         TestHierarchy testHierarchy = new TestHierarchy();
         testHierarchy.setName(jsonTestHierarchy.getName());
         testHierarchy.setRank(jsonTestHierarchy.getRank());
@@ -116,6 +139,7 @@ public class TestHierarchyServiceImpl implements TestHierarchyService {
 
         if (parent == null) {
             testHierarchy.setReference(testHierarchy);
+            testHierarchy.setWebextEngine(webextEngine);
         } else {
             testHierarchy.setReference(parent.getReference());
         }
@@ -124,7 +148,7 @@ public class TestHierarchyServiceImpl implements TestHierarchyService {
 
         Collection<TestHierarchy> children = new ArrayList<>();
         for (JsonTestHierarchy childJson : jsonTestHierarchy.getChildren()) {
-            children.add(importTestHierarchy(childJson, testHierarchy));
+            children.add(importTestHierarchy(childJson, testHierarchy, webextEngine));
         }
         testHierarchy.setChildren(children);
         return testHierarchy;