+++ tanaguru-engine/tanaguru-audit-runner/src/main/java/com/tanaguru/runner/AuditRunnerPage.java
@@ -3,6 +3,8 @@ package com.tanaguru.runner;
 
 import com.tanaguru.domain.entity.audit.Audit;
 import com.tanaguru.domain.entity.audit.TanaguruTest;
+import com.tanaguru.repository.WebextEngineRepository;
+
 import org.openqa.selenium.remote.RemoteWebDriver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -18,15 +20,14 @@ public class AuditRunnerPage extends AbstractAuditRunner {
             Audit audit,
             Collection<String> urls,
             RemoteWebDriver driver,
-            String coreScript,
             long waitTime,
             Collection<Integer> resolutions,
             String basicAuthUrl,
             String basicAuthLogin,
             String basicAuthPassword,
             boolean enableScreenShot,
-            String coreScriptVersion) {
-        super(tanaguruTests, audit, driver, coreScript, waitTime, resolutions, basicAuthUrl, basicAuthLogin, basicAuthPassword, enableScreenShot, coreScriptVersion);
+            WebextEngineRepository webextEngineRepository) {
+        super(tanaguruTests, audit, driver, waitTime, resolutions, basicAuthUrl, basicAuthLogin, basicAuthPassword, enableScreenShot, webextEngineRepository);
         this.urls = urls;
     }