+++ tanaguru-engine/tanaguru-audit-runner/src/main/java/com/tanaguru/runner/AuditRunnerSelenese.java
@@ -4,6 +4,7 @@ package com.tanaguru.runner;
 import com.tanaguru.domain.constant.EAuditLogLevel;
 import com.tanaguru.domain.entity.audit.Audit;
 import com.tanaguru.domain.entity.audit.TanaguruTest;
+import com.tanaguru.repository.WebextEngineRepository;
 import com.tanaguru.selenese.command.SeleneseAudit;
 import com.tanaguru.selenese.command.SeleneseClick;
 import com.tanaguru.selenese.command.SeleneseOpen;
@@ -31,25 +32,23 @@ public class AuditRunnerSelenese extends AbstractAuditRunner {
             Audit audit,
             String scenario,
             RemoteWebDriver driver,
-            String coreScript,
             long waitTime,
             Collection<Integer> resolutions,
             String basicAuthUrl,
             String basicAuthLogin,
             String basicAuthPassword,
             boolean enableScreenShot,
-            String coreScriptVersion) {
+            WebextEngineRepository webextEngineRepository) {
         super(tanaguruTests,
                 audit,
                 driver,
-                coreScript,
                 waitTime,
                 resolutions,
                 basicAuthUrl,
                 basicAuthLogin,
                 basicAuthPassword,
                 enableScreenShot,
-                coreScriptVersion);
+                webextEngineRepository);
         this.scenario = scenario;
     }