+++ tanaguru-engine/tanaguru-audit-runner/src/main/java/com/tanaguru/service/impl/runner/AuditRequestServiceSyncStandaloneImpl.java
@@ -49,8 +49,7 @@ public class AuditRequestServiceSyncStandaloneImpl extends AbstractAuditRunnerSe
             ResultAnalyzerService resultAnalyzerService,
             TestHierarchyRepository testHierarchyRepository,
             ElementResultRepository elementResultRepository,
-            AuditReferenceRepository auditReferenceRepository,
-            WebextEngineRepository webextEngineRepository) {
+            AuditReferenceRepository auditReferenceRepository) {
 
         super(pageRepository,
                 auditRepository,
@@ -61,8 +60,7 @@ public class AuditRequestServiceSyncStandaloneImpl extends AbstractAuditRunnerSe
                 resultAnalyzerService,
                 testHierarchyRepository,
                 elementResultRepository,
-                auditReferenceRepository,
-                webextEngineRepository);
+                auditReferenceRepository);
         this.auditRunnerFactory = auditRunnerFactory;
     }